+++ solr/core/src/java/org/apache/solr/handler/admin/api/CreateAliasAPI.java
@@ -1,0 +1,79 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.solr.handler.admin.api;
+
+import static org.apache.solr.client.solrj.SolrRequest.METHOD.POST;
+import static org.apache.solr.cloud.api.collections.RoutedAlias.CREATE_COLLECTION_PREFIX;
+import static org.apache.solr.common.params.CollectionAdminParams.ROUTER_PREFIX;
+import static org.apache.solr.common.params.CommonParams.ACTION;
+import static org.apache.solr.handler.ClusterAPI.wrapParams;
+import static org.apache.solr.handler.admin.api.CreateCollectionAPI.convertV2CreateCollectionMapToV1ParamMap;
+import static org.apache.solr.handler.api.V2ApiUtils.flattenMapWithPrefix;
+import static org.apache.solr.security.PermissionNameProvider.Name.COLL_EDIT_PERM;
+
+import java.util.HashMap;
+import java.util.Map;
+import org.apache.solr.api.Command;
+import org.apache.solr.api.EndPoint;
+import org.apache.solr.api.PayloadObj;
+import org.apache.solr.client.solrj.request.beans.CreateAliasPayload;
+import org.apache.solr.client.solrj.request.beans.V2ApiConstants;
+import org.apache.solr.common.params.CollectionParams;
+import org.apache.solr.handler.admin.CollectionsHandler;
+
+@EndPoint(
+    path = {"/aliases"},
+    method = POST,
+    permission = COLL_EDIT_PERM)
+public class CreateAliasAPI {
+
+  public static final String V2_CREATE_ALIAS_CMD = "create-alias";
+
+  private final CollectionsHandler collectionsHandler;
+
+  public CreateAliasAPI(CollectionsHandler collectionsHandler) {
+    this.collectionsHandler = collectionsHandler;
+  }
+
+  @Command(name = V2_CREATE_ALIAS_CMD)
+  @SuppressWarnings("unchecked")
+  public void createAlias(PayloadObj<CreateAliasPayload> obj) throws Exception {
+    final CreateAliasPayload v2Body = obj.get();
+    final Map<String, Object> v1Params = v2Body.toMap(new HashMap<>());
+
+    v1Params.put(ACTION, CollectionParams.CollectionAction.CREATEALIAS.toLower());
+    if (v2Body.collections != null && !v2Body.collections.isEmpty()) {
+      final String collectionsStr = String.join(",", v2Body.collections);
+      v1Params.remove(V2ApiConstants.COLLECTIONS);
+      v1Params.put(V2ApiConstants.COLLECTIONS, collectionsStr);
+    }
+    if (v2Body.router != null) {
+      Map<String, Object> routerProperties =
+          (Map<String, Object>) v1Params.remove(V2ApiConstants.ROUTER_KEY);
+      flattenMapWithPrefix(routerProperties, v1Params, ROUTER_PREFIX);
+    }
+    if (v2Body.createCollectionParams != null && !v2Body.createCollectionParams.isEmpty()) {
+      final Map<String, Object> createCollectionMap =
+          (Map<String, Object>) v1Params.remove(V2ApiConstants.CREATE_COLLECTION_KEY);
+      convertV2CreateCollectionMapToV1ParamMap(createCollectionMap);
+      flattenMapWithPrefix(createCollectionMap, v1Params, CREATE_COLLECTION_PREFIX);
+    }
+
+    collectionsHandler.handleRequestBody(wrapParams(obj.getRequest(), v1Params), obj.getResponse());
+  }
+}
