+++ solr/core/src/test/org/apache/solr/cloud/CreateRoutedAliasTest.java
@@ -94,7 +94,7 @@ public class CreateRoutedAliasTest extends SolrCloudTestCase {
 
     final String baseUrl = cluster.getRandomJetty(random()).getBaseUrl().toString();
     // TODO fix Solr test infra so that this /____v2/ becomes /api/
-    HttpPost post = new HttpPost(baseUrl + "/____v2/c");
+    HttpPost post = new HttpPost(baseUrl + "/____v2/aliases");
     post.setEntity(
         new StringEntity(
             "{\n"
