+++ solr/core/src/test/org/apache/solr/handler/admin/TestCollectionAPIs.java
@@ -46,7 +46,6 @@ import org.apache.solr.common.util.Pair;
 import org.apache.solr.common.util.Utils;
 import org.apache.solr.core.CoreContainer;
 import org.apache.solr.handler.ClusterAPI;
-import org.apache.solr.handler.CollectionsAPI;
 import org.apache.solr.request.LocalSolrQueryRequest;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
@@ -81,9 +80,6 @@ public class TestCollectionAPIs extends SolrTestCaseJ4 {
     ApiBag apiBag;
     try (MockCollectionsHandler collectionsHandler = new MockCollectionsHandler()) {
       apiBag = new ApiBag(false);
-      final CollectionsAPI collectionsAPI = new CollectionsAPI(collectionsHandler);
-      apiBag.registerObject(new CollectionsAPI(collectionsHandler));
-      apiBag.registerObject(collectionsAPI.collectionsCommands);
       for (Api api : collectionsHandler.getApis()) {
         apiBag.register(api);
       }
@@ -124,7 +120,7 @@ public class TestCollectionAPIs extends SolrTestCaseJ4 {
 
     compareOutput(
         apiBag,
-        "/collections",
+        "/aliases",
         POST,
         "{create-alias:{name: aliasName , collections:[c1,c2] }}",
         "{operation : createalias, name: aliasName, collections:\"c1,c2\" }");
