+++ solr/core/src/test/org/apache/solr/handler/admin/V2CollectionsAPIMappingTest.java
@@ -31,11 +31,13 @@ import org.apache.solr.common.params.CoreAdminParams;
 import org.apache.solr.common.params.ShardParams;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.core.backup.BackupManager;
-import org.apache.solr.handler.CollectionsAPI;
+import org.apache.solr.handler.admin.api.CreateAliasAPI;
+import org.apache.solr.handler.admin.api.CreateCollectionAPI;
+import org.apache.solr.handler.admin.api.RestoreCollectionAPI;
 import org.junit.Test;
 
 /**
- * Unit tests for the API mappings found in {@link org.apache.solr.handler.CollectionsAPI}.
+ * Unit tests for the "create collection", "create alias" and "restore collection" v2 APIs.
  *
  * <p>This test bears many similarities to {@link TestCollectionAPIs} which appears to test the
  * mappings indirectly by checking message sent to the ZK overseer (which is similar, but not
@@ -52,9 +54,9 @@ public class V2CollectionsAPIMappingTest extends V2ApiMappingTest<CollectionsHan
 
   @Override
   public void populateApiBag() {
-    final CollectionsAPI collectionsAPI = new CollectionsAPI(getRequestHandler());
-    apiBag.registerObject(collectionsAPI);
-    apiBag.registerObject(collectionsAPI.collectionsCommands);
+    apiBag.registerObject(new CreateCollectionAPI(getRequestHandler()));
+    apiBag.registerObject(new CreateAliasAPI(getRequestHandler()));
+    apiBag.registerObject(new RestoreCollectionAPI(getRequestHandler()));
   }
 
   @Override
@@ -117,7 +119,7 @@ public class V2CollectionsAPIMappingTest extends V2ApiMappingTest<CollectionsHan
   public void testCreateAliasAllProperties() throws Exception {
     final SolrParams v1Params =
         captureConvertedV1Params(
-            "/collections",
+            "/aliases",
             "POST",
             "{'create-alias': {"
                 + "'name': 'aliasName', "
@@ -179,10 +181,9 @@ public class V2CollectionsAPIMappingTest extends V2ApiMappingTest<CollectionsHan
   public void testRestoreAllProperties() throws Exception {
     final SolrParams v1Params =
         captureConvertedV1Params(
-            "/collections",
+            "/backups/backupName",
             "POST",
             "{'restore-collection': {"
-                + "'name': 'backupName', "
                 + "'collection': 'collectionName', "
                 + "'location': '/some/location/uri', "
                 + "'repository': 'someRepository', "
