+++ solr/solrj/src/java/org/apache/solr/client/solrj/request/beans/RestoreCollectionPayload.java
@@ -32,9 +32,6 @@ public class RestoreCollectionPayload implements ReflectMapWriter {
   @JsonProperty(required = true)
   public String collection;
 
-  @JsonProperty(required = true)
-  public String name;
-
   @JsonProperty public String location;
 
   @JsonProperty public String repository;
